requirements.txt
Django>=5.0,<6.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
celery>=5.4
redis>=5.0
requests>=2.31
pycryptodome>=3.20
base58>=2.1
safe-pysha3>=1.0
psycopg2-binary>=2.9
websocket-client>=1.7
